[serverless] refactor serverless reporter
