+++ v1/ao/internal/reporter/reporter_test.go
@@ -4,8 +4,6 @@ package reporter
 
 import (
 	"context"
-	"encoding/base64"
-	"encoding/json"
 	"io"
 	"net"
 	"os"
@@ -764,61 +762,8 @@ func TestHttpsProxy(t *testing.T) {
 	testProxy(t, "https://usr:pwd@localhost:12345")
 }
 
-func TestServerless(t *testing.T) {
-	serverless := os.Getenv("APPOPTICS_SERVERLESS")
-	defer func() { os.Setenv("APPOPTICS_SERVERLESS", serverless) }()
-
-	os.Setenv("APPOPTICS_SERVERLESS", "true")
-	config.Load()
-
-	var sb utils.SafeBuffer
-	globalReporter = newServerlessReporter(&sb)
-	r := globalReporter.(*grpcReporter)
-
-	ctx := newTestContext(t)
-	ev1, err := ctx.newEvent(LabelInfo, "layer1")
-	assert.NoError(t, err)
-	assert.NoError(t, r.reportEvent(ctx, ev1))
-
-	ReportSpan(&metrics.HTTPSpanMessage{})
-	IncrementMetric("custom_metric", metrics.MetricOptions{Count: 1})
-
-	assert.Nil(t, r.Flush())
-	arr := strings.Split(strings.TrimRight(sb.String(), "\n"), "\n")
-	evtCnt := 0
-	metricCnt := 0
-	for _, s := range arr {
-		sm := &ServerlessMessage{}
-		assert.Nil(t, json.Unmarshal([]byte(s), sm))
-
-		metricCnt += len(sm.Data.Metrics)
-		evtCnt += len(sm.Data.Events)
-
-		for _, s := range sm.Data.Events {
-			evtByte, err := base64.StdEncoding.DecodeString(s)
-			assert.Nil(t, err)
-			evt := string(evtByte)
-			assert.Contains(t, evt, "layer1")
-		}
-	}
-	assert.Equal(t, 2, metricCnt)
-	assert.Equal(t, 1, evtCnt)
-}
-
-func TestServerlessShutdown(t *testing.T) {
-	serverless := os.Getenv("APPOPTICS_SERVERLESS")
-	defer func() { os.Setenv("APPOPTICS_SERVERLESS", serverless) }()
-
-	os.Setenv("APPOPTICS_SERVERLESS", "true")
-	config.Load()
-
-	globalReporter = newServerlessReporter(os.Stderr)
-	r := globalReporter.(*grpcReporter)
-	assert.Nil(t, r.ShutdownNow())
-}
-
 func TestFlush(t *testing.T) {
 	globalReporter = newGRPCReporter()
 	r := globalReporter.(*grpcReporter)
-	assert.Error(t, r.Flush())
+	assert.NoError(t, r.Flush())
 }
